# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGI Judd's Idea Catalog API",
    "description": "Transform LLM conversations into actionable Linear projects",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/callback": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Auth Callback",
        "description": "Handle Google OAuth callback.\n\nArgs:\n    response: FastAPI response object\n    code: Authorization code from Google\n    state: State parameter for verification\n    error: Error parameter if OAuth failed\n\nReturns:\n    Redirect response to frontend dashboard or login page",
        "operationId": "auth_callback_api_auth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Authorization code from Google",
              "title": "Code"
            },
            "description": "Authorization code from Google"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "State parameter for security",
              "title": "State"
            },
            "description": "State parameter for security"
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Error from OAuth provider",
              "title": "Error"
            },
            "description": "Error from OAuth provider"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/status": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Get Auth Status",
        "description": "Check authentication status.\n\nArgs:\n    request: Current request\n    session_token: Session token from cookie\n\nReturns:\n    Authentication status and user info if authenticated",
        "operationId": "get_auth_status_api_auth_status_get",
        "parameters": [
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Logout",
        "description": "Log out current user.\n\nArgs:\n    response: FastAPI response object\n    session_token: Session token from cookie\n\nReturns:\n    Success message",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Logout Api Auth Logout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{conversation_id}/project-draft/chat": {
      "get": {
        "summary": "Get Chat History",
        "description": "Get chat history for a conversation's project draft.",
        "operationId": "get_chat_history_api_conversations__conversation_id__project_draft_chat_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ChatHistoryResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Get Chat History Api Conversations  Conversation Id  Project Draft Chat Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations": {
      "get": {
        "summary": "List Conversations",
        "description": "Get a paginated list of all imported conversations.",
        "operationId": "list_conversations_api_conversations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConversationListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response List Conversations Api Conversations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthStatus": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated",
            "description": "Whether the user is authenticated"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthUser"
              },
              {
                "type": "null"
              }
            ],
            "description": "User information if authenticated"
          }
        },
        "type": "object",
        "required": [
          "authenticated"
        ],
        "title": "AuthStatus",
        "description": "Authentication status response."
      },
      "AuthUser": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Database user ID"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User email address"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "User display name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "AuthUser",
        "description": "User information returned by authentication endpoints."
      },
...
      "ChatHistoryResponse": {
        "properties": {
          "chat_messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Chat Messages",
            "description": "List of chat messages"
          }
        },
        "type": "object",
        "required": [
          "chat_messages"
        ],
        "title": "ChatHistoryResponse",
        "description": "Chat history response."
      },
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Message role ('user' or 'assistant')"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Message content"
          },
          "sequence_number": {
            "type": "integer",
            "title": "Sequence Number",
            "description": "Sequence number for ordering"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO format creation timestamp"
          },
          "sent_by_user_id": {
            "type": "integer",
            "title": "Sent By User Id",
            "description": "ID of the user who sent this message"
          },
          "sent_by_user_name": {
            "type": "string",
            "title": "Sent By User Name",
            "description": "Name of the user who sent this message"
          },
          "sent_by_user_email": {
            "type": "string",
            "title": "Sent By User Email",
            "description": "Email of the user who sent this message"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/FileAttachment"
            },
            "type": "array",
            "title": "Attachments",
            "description": "File attachments for this message"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "sequence_number",
          "created_at",
          "sent_by_user_id",
          "sent_by_user_name",
          "sent_by_user_email"
        ],
        "title": "ChatMessage",
        "description": "Represents a single chat message."
      },
...
      "ConversationListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "import_date": {
            "type": "string",
            "title": "Import Date"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "is_locked": {
            "type": "boolean",
            "title": "Is Locked"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "user_email": {
            "type": "string",
            "title": "User Email"
          },
          "project_draft_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Draft Title"
          },
          "project_draft_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Draft Description"
          },
          "linear_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linear Url"
          },
          "last_user_message_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last User Message Content"
          },
          "last_assistant_message_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Assistant Message Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "import_date",
          "created_at",
          "updated_at",
          "is_locked",
          "user_id",
          "user_name",
          "user_email"
        ],
        "title": "ConversationListItem",
        "description": "Response model for dashboard list view."
      },
      "ConversationListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationListItem"
            },
            "type": "array",
            "title": "Conversations",
            "description": "List of conversations"
          }
        },
        "type": "object",
        "required": [
          "conversations"
        ],
        "title": "ConversationListResponse",
        "description": "Response for listing conversations."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema."
      },
      "FileAttachment": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Unique attachment ID"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "File size in bytes"
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "MIME type of the file"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key",
            "description": "S3 storage key"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO format creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "file_size",
          "file_type",
          "s3_key",
          "created_at"
        ],
        "title": "FileAttachment",
        "description": "File attachment model for chat messages."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}